kingdomfactory created, base refakt
