+++ main.js
@@ -24,8 +24,9 @@ var historyController = require('./public/history/historyController.js');
 var historyFactory = require('./public/history/historyFactory.js');
 var leaderboardController = require('./public/leaderboard/leaderboardController.js');
 var leaderboardFactory = require('./public/leaderboard/leaderboardFactory.js');
-var mapController = require('./public/map/mapController.js');
+var kingdomFactory = require('./public/map/kingdomFactory.js');
 var mapFactory = require('./public/map/mapFactory.js');
+var mapController = require('./public/map/mapController.js');
 var profileController = require('./public/profile/profileController.js');
 var profileFactory = require('./public/profile/profileFactory.js');
 var troopsController = require('./public/troops/troopsController.js');
