+++ public/map/mapFactory.js
@@ -19,5 +19,4 @@ angular
           }
         )
      }
-
-});
+})
